+++ bdc_stac/routes.py
@@ -8,13 +8,14 @@
 """Routes for the BDC-STAC API."""
 
 import gzip
+from copy import deepcopy
 
 from bdc_auth_client.decorators import oauth2
 from flask import abort, current_app, request, send_from_directory
 from werkzeug.exceptions import HTTPException, InternalServerError
 from werkzeug.urls import url_encode
 
-from .config import BDC_STAC_API_VERSION, BDC_STAC_BASE_URL, BDC_STAC_ID, BDC_STAC_TITLE
+from .config import BDC_STAC_API_VERSION, BDC_STAC_ASSETS_ARGS, BDC_STAC_BASE_URL, BDC_STAC_ID, BDC_STAC_TITLE
 from .data import InvalidBoundingBoxError, get_catalog, get_collection_items, get_collections, make_geojson, session
 
 BASE_URL = BDC_STAC_BASE_URL
@@ -30,6 +31,8 @@ def teardown_appcontext(exceptions=None):
 def after_request(response):
     """Enable CORS and compress response."""
     response.headers.add("Access-Control-Allow-Origin", "*")
+    response.headers.add("Access-Control-Allow-Headers", "Content-Type")
+    response.headers.add("Access-Control-Allow-Methods", "GET, POST")
 
     accept_encoding = request.headers.get("Accept-Encoding", "")
 
@@ -69,7 +72,13 @@ def conformance():
 @oauth2(required=False, throw_exception=False)
 def index(roles=[], access_token=""):
     """Landing page of this API."""
-    access_token = f"?access_token={access_token}" if access_token else ""
+    assets_kwargs = None
+
+    if BDC_STAC_ASSETS_ARGS:
+        assets_kwargs = {arg: request.args.get(arg) for arg in BDC_STAC_ASSETS_ARGS.split(",")}
+        if access_token:
+            assets_kwargs["access_token"] = access_token
+        assets_kwargs = "?" + url_encode(assets_kwargs) if url_encode(assets_kwargs) else ""
 
     collections = get_catalog(roles=roles)
     catalog = dict()
@@ -98,7 +107,7 @@ def index(roles=[], access_token=""):
     for collection in collections:
         links.append(
             {
-                "href": f"{BASE_URL}/collections/{collection.name}{access_token}",
+                "href": f"{BASE_URL}/collections/{collection.name}{assets_kwargs}",
                 "rel": "child",
                 "type": "application/json",
                 "title": collection.title,
@@ -114,7 +123,13 @@ def index(roles=[], access_token=""):
 @oauth2(required=False, throw_exception=False)
 def root(roles=[], access_token=""):
     """Return the root catalog or collection."""
-    access_token = f"?access_token={access_token}" if access_token else ""
+    assets_kwargs = None
+
+    if BDC_STAC_ASSETS_ARGS:
+        assets_kwargs = {arg: request.args.get(arg) for arg in BDC_STAC_ASSETS_ARGS.split(",")}
+        if access_token:
+            assets_kwargs["access_token"] = access_token
+        assets_kwargs = "?" + url_encode(assets_kwargs) if url_encode(assets_kwargs) else ""
 
     collections = get_collections(roles=roles)
     response = dict()
@@ -122,25 +137,25 @@ def root(roles=[], access_token=""):
     for collection in collections:
         links = [
             {
-                "href": f"{BASE_URL}/collections/{collection['id']}{access_token}",
+                "href": f"{BASE_URL}/collections/{collection['id']}{assets_kwargs}",
                 "rel": "self",
                 "type": "application/json",
                 "title": "Link to this document",
             },
             {
-                "href": f"{BASE_URL}/collections/{collection['id']}/items{access_token}",
+                "href": f"{BASE_URL}/collections/{collection['id']}/items{assets_kwargs}",
                 "rel": "items",
                 "type": "application/json",
                 "title": f"Items of the collection {collection['id']}",
             },
             {
-                "href": f"{BASE_URL}/collections{access_token}",
+                "href": f"{BASE_URL}/collections{assets_kwargs}",
                 "rel": "parent",
                 "type": "application/json",
                 "title": "Link to catalog collections",
             },
             {
-                "href": f"{BASE_URL}/{access_token}",
+                "href": f"{BASE_URL}/{assets_kwargs}",
                 "rel": "root",
                 "type": "application/json",
                 "title": "API landing page (root catalog)",
@@ -160,7 +175,13 @@ def collections_id(collection_id, roles=[], access_token=""):
 
     :param collection_id: identifier (name) of a specific collection
     """
-    access_token = f"?access_token={access_token}" if access_token else ""
+    assets_kwargs = None
+
+    if BDC_STAC_ASSETS_ARGS:
+        assets_kwargs = {arg: request.args.get(arg) for arg in BDC_STAC_ASSETS_ARGS.split(",")}
+        if access_token:
+            assets_kwargs["access_token"] = access_token
+        assets_kwargs = "?" + url_encode(assets_kwargs) if url_encode(assets_kwargs) else ""
 
     collection = get_collections(collection_id, roles=roles)
 
@@ -171,25 +192,25 @@ def collections_id(collection_id, roles=[], access_token=""):
 
     links = [
         {
-            "href": f"{BASE_URL}/collections/{collection['id']}{access_token}",
+            "href": f"{BASE_URL}/collections/{collection['id']}{assets_kwargs}",
             "rel": "self",
             "type": "application/json",
             "title": "Link to this document",
         },
         {
-            "href": f"{BASE_URL}/collections/{collection['id']}/items{access_token}",
+            "href": f"{BASE_URL}/collections/{collection['id']}/items{assets_kwargs}",
             "rel": "items",
             "type": "application/json",
             "title": f"Items of the collection {collection['id']}",
         },
         {
-            "href": f"{BASE_URL}/collections{access_token}",
+            "href": f"{BASE_URL}/collections{assets_kwargs}",
             "rel": "parent",
             "type": "application/json",
             "title": "Link to catalog collections",
         },
         {
-            "href": f"{BASE_URL}/{access_token}",
+            "href": f"{BASE_URL}/{assets_kwargs}",
             "rel": "root",
             "type": "application/json",
             "title": "API landing page (root catalog)",
@@ -208,8 +229,6 @@ def collection_items(collection_id, roles=[], access_token=""):
 
     :param collection_id: identifier (name) of a specific collection
     """
-    access_token = f"?access_token={access_token}" if access_token else ""
-
     items = get_collection_items(collection_id=collection_id, roles=roles, **request.args.to_dict())
 
     links = [
@@ -239,7 +258,15 @@ def collection_items(collection_id, roles=[], access_token=""):
     gjson["stac_extensions"] = ["checksum", "commons", "context", "eo"]
     gjson["type"] = "FeatureCollection"
 
-    features = make_geojson(items.items, links, access_token=access_token)
+    assets_kwargs = None
+
+    if BDC_STAC_ASSETS_ARGS:
+        assets_kwargs = {arg: request.args.get(arg) for arg in BDC_STAC_ASSETS_ARGS.split(",")}
+        if access_token:
+            assets_kwargs["access_token"] = access_token
+        assets_kwargs = "?" + url_encode(assets_kwargs) if url_encode(assets_kwargs) else ""
+
+    features = make_geojson(items.items, links, assets_kwargs=assets_kwargs)
 
     gjson["links"] = []
 
@@ -250,15 +277,22 @@ def collection_items(collection_id, roles=[], access_token=""):
     gjson["context"] = context
 
     args = request.args.copy()
+
     if items.has_next:
         args["page"] = items.next_num
         gjson["links"].append(
-            {"href": f"{BASE_URL}/collections/{collection_id}/items?" + url_encode(args), "rel": "next"}
+            {
+                "href": f"{BASE_URL}/collections/{collection_id}/items{f'?{url_encode(args)}' if len(args) > 0 else ''}",
+                "rel": "next",
+            }
         )
     if items.has_prev:
         args["page"] = items.prev_num
         gjson["links"].append(
-            {"href": f"{BASE_URL}/collections/{collection_id}/items?" + url_encode(args), "rel": "prev"}
+            {
+                "href": f"{BASE_URL}/collections/{collection_id}/items{f'?{url_encode(args)}' if len(args) > 0 else ''}",
+                "rel": "prev",
+            }
         )
 
     gjson["features"] = features
@@ -273,8 +307,6 @@ def items_id(collection_id, item_id, roles=[], access_token=""):
     :param collection_id: identifier (name) of a specific collection
     :param item_id: identifier (name) of a specific item
     """
-    access_token = f"?access_token={access_token}" if access_token else ""
-
     item = get_collection_items(collection_id=collection_id, roles=roles, item_id=item_id)
     links = [
         {"href": f"{BASE_URL}/collections/", "rel": "self"},
@@ -283,7 +315,15 @@ def items_id(collection_id, item_id, roles=[], access_token=""):
         {"href": f"{BASE_URL}/", "rel": "root"},
     ]
 
-    gjson = make_geojson(item.items, links, access_token=access_token)
+    assets_kwargs = None
+
+    if BDC_STAC_ASSETS_ARGS:
+        assets_kwargs = {arg: request.args.get(arg) for arg in BDC_STAC_ASSETS_ARGS.split(",")}
+        if access_token:
+            assets_kwargs["access_token"] = access_token
+        assets_kwargs = "?" + url_encode(assets_kwargs) if url_encode(assets_kwargs) else ""
+
+    gjson = make_geojson(item.items, links, assets_kwargs=assets_kwargs)
 
     if len(gjson) > 0:
         return gjson[0]
@@ -295,8 +335,6 @@ def items_id(collection_id, item_id, roles=[], access_token=""):
 @oauth2(required=False, throw_exception=False)
 def stac_search(roles=[], access_token=""):
     """Search STAC items with simple filtering."""
-    access_token = f"?access_token={access_token}" if access_token else ""
-
     bbox, datetime, ids, collections, page, limit, intersects, query = None, None, None, None, None, None, None, None
     if request.method == "POST":
         if request.is_json:
@@ -353,7 +391,15 @@ def stac_search(roles=[], access_token=""):
     gjson = dict()
     gjson["type"] = "FeatureCollection"
 
-    features = make_geojson(items.items, links, access_token=access_token)
+    assets_kwargs = None
+
+    if BDC_STAC_ASSETS_ARGS:
+        assets_kwargs = {arg: request.args.get(arg) for arg in BDC_STAC_ASSETS_ARGS.split(",")}
+        if access_token:
+            assets_kwargs["access_token"] = access_token
+        assets_kwargs = "?" + url_encode(assets_kwargs) if url_encode(assets_kwargs) else ""
+
+    features = make_geojson(items.items, links, assets_kwargs=assets_kwargs)
 
     gjson["links"] = []
     context = dict()
@@ -362,12 +408,37 @@ def stac_search(roles=[], access_token=""):
     gjson["context"] = context
 
     args = request.args.copy()
+
     if items.has_next:
-        args["page"] = items.next_num
-        gjson["links"].append({"href": f"{BASE_URL}/search?" + url_encode(args), "rel": "next"})
+        if request.method == "GET":
+            args["page"] = items.next_num
+
+        next_links = {
+            "href": f"{BASE_URL}/search{f'?{url_encode(args)}' if len(args) > 0 else ''}",
+            "rel": "next",
+        }
+
+        if request.method == "POST":
+            next_links["body"] = deepcopy(request_json)
+            next_links["body"]["page"] = items.next_num
+            next_links["method"] = "POST"
+            next_links["merge"] = True
+        gjson["links"].append(next_links)
     if items.has_prev:
-        args["page"] = items.prev_num
-        gjson["links"].append({"href": f"{BASE_URL}/search?" + url_encode(args), "rel": "prev"})
+        if request.method == "GET":
+            args["page"] = items.prev_num
+
+        prev_links = {
+            "href": f"{BASE_URL}/search{f'?{url_encode(args)}' if len(args) > 0 else ''}",
+            "rel": "prev",
+        }
+
+        if request.method == "POST":
+            prev_links["body"] = deepcopy(request_json)
+            prev_links["body"]["page"] = items.prev_num
+            prev_links["method"] = "POST"
+            prev_links["merge"] = True
+        gjson["links"].append(prev_links)
 
     gjson["features"] = features
 
@@ -376,6 +447,7 @@ def stac_search(roles=[], access_token=""):
 
 @current_app.route("/schemas/<string:schema_name>")
 def list_schema(schema_name):
+    """Return jsonschemas."""
     return send_from_directory("spec/jsonschemas", schema_name)