+++ CHANGES.rst
@@ -10,8 +10,25 @@
 Changes
 =======
 
-Version 0.9.0-12
-----------------
+
+
+Version 0.9.0-14 (2021-04-01)
+----------------------------
+
+- Improve pagination when using POST in /search (`#141 <https://github.com/brazil-data-cube/bdc-stac/pull/141>`_)
+- remove the question mark if there are no args in the url (`#147 <https://github.com/brazil-data-cube/bdc-stac/pull/147>`_)
+
+Version 0.9.0-13 (2021-01-28)
+----------------------------
+
+- Add drone support (`#133 <https://github.com/brazil-data-cube/bdc-stac/issues/133>`_)
+- Remove collection metadata from item (`#136 <https://github.com/brazil-data-cube/bdc-stac/issues/136>`_)
+- Add missing CORS headers (`#138 <https://github.com/brazil-data-cube/bdc-stac/issues/138>`_)
+- Pass auth parameters to asset url (`#130 <https://github.com/brazil-data-cube/bdc-stac/issues/130>`_)
+- Fix bug when access collection item id directly (`#140 <https://github.com/brazil-data-cube/bdc-stac/issues/140>`_)
+
+Version 0.9.0-12 (2021-01-14)
+----------------------------
 
 - Add configuration for catalog description and id. (`#129 <https://github.com/brazil-data-cube/bdc-stac/issues/129>`_)